fifth commit

+++ app/src/main/java/com/kozin/weatherkotlin/data/response/future/FutureList.kt
@@ -2,13 +2,13 @@ package com.kozin.weatherkotlin.data.response.future
 
 data class FutureList(
     val clouds: Clouds,
-    val dt: Int,
+    val dt: Double,
     val dt_txt: String,
     val main: Main,
-    val pop: Int,
+    val pop: Double,
     val rain: Rain,
     val sys: Sys,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<Weather>,
     val wind: Wind
 )
