+++ app/src/main/java/com/kozin/weatherkotlin/data/response/future/FutureWeatherEntry.kt
@@ -2,8 +2,8 @@ package com.kozin.weatherkotlin.data.response.future
 
 data class FutureWeatherEntry(
     val city: City,
-    val cnt: Int,
+    val cnt: Double,
     val cod: String,
     val list: List<FutureList>,
-    val message: Int
+    val message: Double
 )
