+++ app/src/main/java/com/kozin/weatherkotlin/data/response/future/Main.kt
@@ -2,10 +2,10 @@ package com.kozin.weatherkotlin.data.response.future
 
 data class Main(
     val feels_like: Double,
-    val grnd_level: Int,
-    val humidity: Int,
-    val pressure: Int,
-    val sea_level: Int,
+    val grnd_level: Double,
+    val humidity: Double,
+    val pressure: Double,
+    val sea_level: Double,
     val temp: Double,
     val temp_kf: Double,
     val temp_max: Double,
