+++ app/src/main/java/com/kozin/weatherkotlin/data/response/future/Weather.kt
@@ -3,6 +3,6 @@ package com.kozin.weatherkotlin.data.response.future
 data class Weather(
     val description: String,
     val icon: String,
-    val id: Int,
+    val id: Double,
     val main: String
 )
